objects/switch: merge switch collision in TRX
This merges switch collision in TRX, along with animated interactions
for TR2.

Resolves #4067.
