+++ src/libtrx/game/objects/general/switch.c
@@ -3,7 +3,13 @@
 #include "config.h"
 #include "game/const.h"
 #include "game/items.h"
+#include "game/lara.h"
 #include "game/objects.h"
+
+typedef struct {
+    XYZ_32 normal;
+    XYZ_32 controlled;
+} M_SWITCH_POS;
 
 static const OBJECT_BOUNDS m_SwitchBounds = {
     .shift = {
@@ -16,10 +22,10 @@ static const OBJECT_BOUNDS m_SwitchBounds = {
     },
 };
 
-static OBJECT_BOUNDS m_SwitchBoundsControlled = {
+static const OBJECT_BOUNDS m_SwitchBoundsControlled = {
     .shift = {
-        .min = { .x = +0, .y = +0, .z = +0, },
-        .max = { .x = +0, .y = +0, .z = +0, },
+        .min = { .x = -WALL_L / 2, .y = +0, .z = -200, },
+        .max = { .x = +WALL_L / 2, .y = +0, .z = +200, },
     },
     .rot = {
         .min = { .x = -10 * DEG_1, .y = -30 * DEG_1, .z = -10 * DEG_1, },
@@ -38,14 +44,32 @@ static const OBJECT_BOUNDS m_SwitchBoundsUW = {
     },
 };
 
+static const XYZ_32 m_SwitchUWPosition = { .x = 0, .y = 0, .z = 108 };
+
+static const M_SWITCH_POS m_SmallSwitchPosition = {
+    .normal = { .x = 0, .y = 0, .z = 362 },
+    .controlled = { .x = 0, .y = 0, .z = 80 },
+};
+
+static const M_SWITCH_POS m_PushSwitchPosition = {
+    .normal = { .x = 0, .y = 0, .z = 292 },
+    .controlled = { .x = 0, .y = 0, .z = 146 },
+};
+
+static const M_SWITCH_POS m_WallSwitchPosition = {
+    .normal = { .x = 0, .y = 0, .z = 128 },
+    .controlled = { .x = 0, .y = 0, .z = 64 },
+};
+
+static const M_SWITCH_POS m_AirlockPosition = {
+    .normal = { .x = 0, .y = 0, .z = 212 },
+    .controlled = { .x = 0, .y = 0, .z = 106 },
+};
+
 static const OBJECT_BOUNDS *M_Bounds(void)
 {
-#if TR_VERSION == 1
-    if (g_Config.gameplay.enable_walk_to_items) {
-        return &m_SwitchBoundsControlled;
-    }
-#endif
-    return &m_SwitchBounds;
+    return g_Config.gameplay.enable_walk_to_items ? &m_SwitchBoundsControlled
+                                                  : &m_SwitchBounds;
 }
 
 static const OBJECT_BOUNDS *M_BoundsUW(void)
@@ -64,6 +88,244 @@ static void M_Control(const int16_t item_num)
     Item_Animate(item);
 }
 
+static void M_AlignLara(ITEM *const lara_item, ITEM *const switch_item)
+{
+    lara_item->rot.y = switch_item->rot.y;
+    switch (switch_item->object_id) {
+    case O_SWITCH_TYPE_AIRLOCK:
+        Lara_AlignPosition(switch_item, &m_AirlockPosition.normal);
+        break;
+
+    case O_SWITCH_TYPE_SMALL:
+        Lara_AlignPosition(switch_item, &m_SmallSwitchPosition.normal);
+        break;
+
+    case O_SWITCH_TYPE_BUTTON:
+        Lara_AlignPosition(switch_item, &m_PushSwitchPosition.normal);
+        break;
+
+    default:
+        break;
+    }
+}
+
+static bool M_MoveLaraControlled(
+    const ITEM *const item, const BOUNDS_16 *const bounds)
+{
+    XYZ_32 shift;
+    switch (item->object_id) {
+    case O_SWITCH_TYPE_AIRLOCK:
+        shift = m_AirlockPosition.controlled;
+        break;
+    case O_SWITCH_TYPE_SMALL:
+        shift = m_SmallSwitchPosition.controlled;
+        break;
+    case O_SWITCH_TYPE_BUTTON:
+        shift = m_PushSwitchPosition.controlled;
+        break;
+    default:
+        shift = m_WallSwitchPosition.controlled;
+        break;
+    }
+
+    const XYZ_32 move_vector = {
+        .x = 0,
+        .y = 0,
+        .z = bounds->min.z - shift.z,
+    };
+    return Lara_MovePosition(item, &move_vector);
+}
+
+static void M_SwitchOn(ITEM *const switch_item, ITEM *const lara_item)
+{
+    switch (switch_item->object_id) {
+    case O_SWITCH_TYPE_SMALL:
+        Item_SwitchToAnim(lara_item, LA(LA_SWITCH_SMALL_DOWN), 0);
+        break;
+
+    case O_SWITCH_TYPE_BUTTON:
+        Item_SwitchToAnim(lara_item, LA(LA_BUTTON_PUSH), 0);
+        break;
+
+    default:
+        Item_SwitchToAnim(lara_item, LA(LA_WALL_SWITCH_DOWN), 0);
+        break;
+    }
+
+    lara_item->current_anim_state = LS(LS_SWITCH_ON);
+    switch_item->goal_anim_state = SWITCH_STATE_OFF;
+}
+
+static void M_SwitchOff(ITEM *const switch_item, ITEM *const lara_item)
+{
+    lara_item->current_anim_state = LS(LS_SWITCH_OFF);
+
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    switch (switch_item->object_id) {
+    case O_SWITCH_TYPE_AIRLOCK:
+#if TR_VERSION >= 2
+        Item_SwitchToObjAnim(lara_item, LS_EXTRA_BREATH, 0, O_LARA_EXTRA);
+        lara_item->current_anim_state = LS_EXTRA_BREATH;
+        lara_item->goal_anim_state = LS_EXTRA_AIRLOCK;
+        Item_Animate(lara_item);
+        lara->extra_anim = true;
+        lara->hit_direction = -1;
+#endif
+        break;
+
+    case O_SWITCH_TYPE_SMALL:
+        Item_SwitchToAnim(lara_item, LA(LA_SWITCH_SMALL_UP), 0);
+        break;
+
+    case O_SWITCH_TYPE_BUTTON:
+        Item_SwitchToAnim(lara_item, LA(LA_BUTTON_PUSH), 0);
+        break;
+
+    default:
+        Item_SwitchToAnim(lara_item, LA(LA_WALL_SWITCH_UP), 0);
+        break;
+    }
+
+    switch_item->goal_anim_state = SWITCH_STATE_ON;
+}
+
+static void M_CollisionControlled(
+    const int16_t item_num, ITEM *const lara_item, COLL_INFO *const coll)
+{
+    ITEM *const item = Item_Get(item_num);
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+
+    if ((g_Input.action && lara->gun_status == LGS_ARMLESS
+         && !lara_item->gravity && lara_item->current_anim_state == LS(LS_STOP)
+         && item->status == IS_INACTIVE)
+        || (lara->interact_target.is_moving
+            && lara->interact_target.item_num == item_num)) {
+        const BOUNDS_16 *const bounds = Item_GetBoundsAccurate(item);
+
+        OBJECT_BOUNDS col_bounds = *Object_Get(item->object_id)->bounds_func();
+        col_bounds.shift.min.x += bounds->min.x;
+        col_bounds.shift.max.x += bounds->max.x;
+        col_bounds.shift.min.z += bounds->min.z;
+        col_bounds.shift.max.z += bounds->max.z;
+
+        if (Lara_TestPosition(item, &col_bounds)) {
+            if (M_MoveLaraControlled(item, bounds)) {
+                if (item->current_anim_state == SWITCH_STATE_ON) {
+                    M_SwitchOn(item, lara_item);
+                } else {
+                    M_SwitchOff(item, lara_item);
+                }
+                lara->head_rot.x = 0;
+                lara->head_rot.y = 0;
+                lara->torso_rot.x = 0;
+                lara->torso_rot.y = 0;
+                lara->interact_target.is_moving = false;
+                lara->interact_target.item_num = NO_ITEM;
+                lara->gun_status = LGS_HANDS_BUSY;
+                Item_AddActive(item_num);
+                item->status = IS_ACTIVE;
+                Item_Animate(item);
+            } else {
+                lara->interact_target.item_num = item_num;
+            }
+        } else if (
+            lara->interact_target.is_moving
+            && lara->interact_target.item_num == item_num) {
+            lara->interact_target.is_moving = false;
+            lara->gun_status = LGS_ARMLESS;
+        }
+    } else if (
+        lara_item->current_anim_state != LS(LS_SWITCH_ON)
+        && lara_item->current_anim_state != LS(LS_SWITCH_OFF)) {
+        Object_Collision(item_num, lara_item, coll);
+    }
+}
+
+static void M_Collision(
+    const int16_t item_num, ITEM *const lara_item, COLL_INFO *const coll)
+{
+    if (g_Config.gameplay.enable_walk_to_items) {
+        M_CollisionControlled(item_num, lara_item, coll);
+        return;
+    }
+
+    ITEM *const item = Item_Get(item_num);
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    const OBJECT *const obj = Object_Get(item->object_id);
+
+    if (!g_Input.action || item->status != IS_INACTIVE
+        || lara->gun_status != LGS_ARMLESS || lara_item->gravity
+        || lara_item->current_anim_state != LS(LS_STOP)
+        || !Lara_TestPosition(item, obj->bounds_func())) {
+        return;
+    }
+
+    if (item->object_id == O_SWITCH_TYPE_AIRLOCK
+        && item->current_anim_state == SWITCH_STATE_ON) {
+        return;
+    }
+
+    M_AlignLara(lara_item, item);
+
+    if (item->current_anim_state == SWITCH_STATE_ON) {
+        M_SwitchOn(item, lara_item);
+    } else {
+        M_SwitchOff(item, lara_item);
+    }
+
+    if (!lara->extra_anim) {
+        lara_item->goal_anim_state = LS(LS_STOP);
+    }
+    lara->gun_status = LGS_HANDS_BUSY;
+
+    item->status = IS_ACTIVE;
+    Item_AddActive(item_num);
+    Item_Animate(item);
+}
+
+static void M_CollisionUW(
+    const int16_t item_num, ITEM *const lara_item, COLL_INFO *const coll)
+{
+    ITEM *const item = Item_Get(item_num);
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    const OBJECT *const obj = Object_Get(item->object_id);
+
+    if (!g_Input.action || item->status != IS_INACTIVE
+        || (lara->water_status != LWS_UNDERWATER
+            && lara->water_status != LWS_CHEAT)
+        || lara->gun_status != LGS_ARMLESS
+        || lara_item->current_anim_state != LS(LS_TREAD)) {
+        return;
+    }
+
+    if (!Lara_TestPosition(item, obj->bounds_func())) {
+        return;
+    }
+
+    if (item->current_anim_state != SWITCH_STATE_OFF
+        && item->current_anim_state != SWITCH_STATE_ON) {
+        return;
+    }
+
+    if (!Lara_MovePosition(item, &m_SwitchUWPosition)) {
+        return;
+    }
+
+    lara_item->fall_speed = 0;
+    Lara_AnimateUntil(lara_item, LS(LS_SWITCH_ON));
+    lara_item->goal_anim_state = LS(LS_TREAD);
+    lara->gun_status = LGS_HANDS_BUSY;
+
+    if (item->current_anim_state == SWITCH_STATE_ON) {
+        item->goal_anim_state = SWITCH_STATE_OFF;
+    } else {
+        item->goal_anim_state = SWITCH_STATE_ON;
+    }
+    item->status = IS_ACTIVE;
+    Item_AddActive(item_num);
+    Item_Animate(item);
+}
+
 static void M_SetupBase(OBJECT *const obj)
 {
     obj->control_func = M_Control;
@@ -74,7 +336,7 @@ static void M_SetupBase(OBJECT *const obj)
 static void M_Setup(OBJECT *const obj)
 {
     M_SetupBase(obj);
-    obj->collision_func = Switch_Collision;
+    obj->collision_func = M_Collision;
     obj->bounds_func = M_Bounds;
 }
 
@@ -88,7 +350,7 @@ static void M_SetupPushButton(OBJECT *const obj)
 static void M_SetupUW(OBJECT *const obj)
 {
     M_SetupBase(obj);
-    obj->collision_func = Switch_CollisionUW;
+    obj->collision_func = M_CollisionUW;
     obj->bounds_func = M_BoundsUW;
 }
