+++ src/tr1/meson.build
@@ -134,7 +134,6 @@ sources = [
   'game/objects/creatures/cutscene_player.c',
   'game/objects/general/earthquake.c',
   'game/objects/general/save_crystal.c',
-  'game/objects/general/switch.c',
   'game/objects/setup.c',
   'game/objects/traps/damocles_sword.c',
   'game/objects/traps/lightning_emitter.c',
