--- src/tr2/game/objects/general/switch.c
@@ -1,165 +1,0 @@
-#include "game/lara.h"
-
-#include <libtrx/game/input.h>
-#include <libtrx/game/lara.h>
-#include <libtrx/game/objects/general/switch.h>
-
-static XYZ_32 g_SmallSwitchPosition = { .x = 0, .y = 0, .z = 362 };
-static XYZ_32 g_PushSwitchPosition = { .x = 0, .y = 0, .z = 292 };
-static XYZ_32 m_AirlockPosition = { .x = 0, .y = 0, .z = 212 };
-static XYZ_32 m_SwitchUWPosition = { .x = 0, .y = 0, .z = 108 };
-
-static void M_AlignLara(ITEM *const lara_item, ITEM *const switch_item)
-{
-    lara_item->rot.y = switch_item->rot.y;
-    switch (switch_item->object_id) {
-    case O_SWITCH_TYPE_AIRLOCK:
-        Lara_AlignPosition(switch_item, &m_AirlockPosition);
-        break;
-
-    case O_SWITCH_TYPE_SMALL:
-        Lara_AlignPosition(switch_item, &g_SmallSwitchPosition);
-        break;
-
-    case O_SWITCH_TYPE_BUTTON:
-        Lara_AlignPosition(switch_item, &g_PushSwitchPosition);
-        break;
-
-    default:
-        break;
-    }
-}
-
-static void M_SwitchOn(ITEM *const switch_item, ITEM *const lara_item)
-{
-    switch (switch_item->object_id) {
-    case O_SWITCH_TYPE_SMALL:
-        Item_SwitchToAnim(lara_item, LA(LA_SWITCH_SMALL_DOWN), 0);
-        break;
-
-    case O_SWITCH_TYPE_BUTTON:
-        Item_SwitchToAnim(lara_item, LA(LA_BUTTON_PUSH), 0);
-        break;
-
-    default:
-        Item_SwitchToAnim(lara_item, LA(LA_WALL_SWITCH_DOWN), 0);
-        break;
-    }
-
-    lara_item->current_anim_state = LS(LS_SWITCH_ON);
-    switch_item->goal_anim_state = SWITCH_STATE_OFF;
-}
-
-static void M_SwitchOff(ITEM *const switch_item, ITEM *const lara_item)
-{
-    lara_item->current_anim_state = LS(LS_SWITCH_OFF);
-
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    switch (switch_item->object_id) {
-    case O_SWITCH_TYPE_AIRLOCK:
-        Item_SwitchToObjAnim(lara_item, LS_EXTRA_BREATH, 0, O_LARA_EXTRA);
-        lara_item->current_anim_state = LS_EXTRA_BREATH;
-        lara_item->goal_anim_state = LS_EXTRA_AIRLOCK;
-        Item_Animate(lara_item);
-        lara->extra_anim = true;
-        lara->hit_direction = -1;
-        break;
-
-    case O_SWITCH_TYPE_SMALL:
-        Item_SwitchToAnim(lara_item, LA(LA_SWITCH_SMALL_UP), 0);
-        break;
-
-    case O_SWITCH_TYPE_BUTTON:
-        Item_SwitchToAnim(lara_item, LA(LA_BUTTON_PUSH), 0);
-        break;
-
-    default:
-        Item_SwitchToAnim(lara_item, LA(LA_WALL_SWITCH_UP), 0);
-        break;
-    }
-
-    switch_item->goal_anim_state = SWITCH_STATE_ON;
-}
-
-void Switch_Collision(
-    const int16_t item_num, ITEM *const lara_item, COLL_INFO *const coll)
-{
-    ITEM *const item = Item_Get(item_num);
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    const OBJECT *const obj = Object_Get(item->object_id);
-
-    if (!g_Input.action || item->status != IS_INACTIVE
-        || lara->gun_status != LGS_ARMLESS || lara_item->gravity
-        || lara_item->current_anim_state != LS(LS_STOP)
-        || !Lara_TestPosition(item, obj->bounds_func())) {
-        return;
-    }
-
-    if (item->object_id == O_SWITCH_TYPE_AIRLOCK
-        && item->current_anim_state == SWITCH_STATE_ON) {
-        return;
-    }
-
-    M_AlignLara(lara_item, item);
-
-    if (item->current_anim_state == SWITCH_STATE_ON) {
-        M_SwitchOn(item, lara_item);
-    } else {
-        M_SwitchOff(item, lara_item);
-    }
-
-    if (!lara->extra_anim) {
-        lara_item->goal_anim_state = LS(LS_STOP);
-    }
-    lara->gun_status = LGS_HANDS_BUSY;
-
-    item->status = IS_ACTIVE;
-    Item_AddActive(item_num);
-    Item_Animate(item);
-}
-
-void Switch_CollisionUW(
-    const int16_t item_num, ITEM *const lara_item, COLL_INFO *const coll)
-{
-    ITEM *const item = Item_Get(item_num);
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    const OBJECT *const obj = Object_Get(item->object_id);
-
-    if (!g_Input.action || item->status != IS_INACTIVE
-        || (lara->water_status != LWS_UNDERWATER
-            && lara->water_status != LWS_CHEAT)
-        || lara->gun_status != LGS_ARMLESS
-        || lara_item->current_anim_state != LS(LS_TREAD)) {
-        return;
-    }
-
-    if (!Lara_TestPosition(item, obj->bounds_func())) {
-        return;
-    }
-
-    if (item->current_anim_state != SWITCH_STATE_OFF
-        && item->current_anim_state != SWITCH_STATE_ON) {
-        return;
-    }
-
-    if (!Lara_MovePosition(item, &m_SwitchUWPosition)) {
-        return;
-    }
-
-    lara_item->fall_speed = 0;
-    lara_item->goal_anim_state = LS(LS_SWITCH_ON);
-    do {
-        Lara_Animate(lara_item);
-    } while (lara_item->current_anim_state != LS(LS_SWITCH_ON));
-    lara_item->goal_anim_state = LS(LS_TREAD);
-    lara->gun_status = LGS_HANDS_BUSY;
-
-    if (item->current_anim_state == SWITCH_STATE_ON) {
-        item->goal_anim_state = SWITCH_STATE_OFF;
-    } else {
-        item->goal_anim_state = SWITCH_STATE_ON;
-    }
-    item->status = IS_ACTIVE;
-    Item_AddActive(item_num);
-    Item_Animate(item);
-}
